ORIS entries: complete logic to push all entries to ORIS

+++ race_regs_all_exc.php
@@ -30,7 +30,7 @@ db_Connect();
 
 $sub_query = (IsLoggedRegistrator() || IsLoggedManager()) ? '' : ' AND '.TBL_USER.'.chief_id = '.$usr->user_id.' OR '.TBL_USER.'.id = '.$usr->user_id;
 
-$query = 'SELECT '.TBL_USER.'.id, kat, termin FROM '.TBL_USER.' LEFT JOIN '.TBL_ZAVXUS.' ON '.TBL_USER.'.id = '.TBL_ZAVXUS.'.id_user AND '.TBL_ZAVXUS.'.id_zavod='.$id.' WHERE '.TBL_USER.'.hidden = 0'.$sub_query;
+$query = 'SELECT '.TBL_USER.'.id, kat, termin, sync_status FROM '.TBL_USER.' LEFT JOIN '.TBL_ZAVXUS.' ON '.TBL_USER.'.id = '.TBL_ZAVXUS.'.id_user AND '.TBL_ZAVXUS.'.id_zavod='.$id.' WHERE '.TBL_USER.'.hidden = 0'.$sub_query;
 
 @$vysledek=query_db($query);
 
@@ -83,13 +83,21 @@ while ($zaznamZ=mysqli_fetch_array($vysledek))
 			if ($kat == "")
 			{	// del
 //				echo "DEL";
-				$result=query_db("DELETE FROM ".TBL_ZAVXUS." WHERE id_zavod = '$id' AND id_user = '$user'")
-					or die("Chyba při provádění dotazu do databáze.");
+				$has_ext_id = !empty($zaznam_z['ext_id']);
+				$is_pending_create = ($zaznamZ['sync_status'] === 'PENDING_CREATE');
+
+				if ($has_ext_id && !$is_pending_create) {
+					$result=query_db("UPDATE ".TBL_ZAVXUS." SET sync_status='PENDING_DELETE' WHERE id_zavod = '$id' AND id_user = '$user'")
+						or die("Chyba při provádění dotazu do databáze.");
+				} else {
+					$result=query_db("DELETE FROM ".TBL_ZAVXUS." WHERE id_zavod = '$id' AND id_user = '$user'")
+						or die("Chyba při provádění dotazu do databáze.");
+					if ($result !== false && mysqli_affected_rows($db_conn) > 0) {
+						query_db("UPDATE ".TBL_RACE." SET prihlasenych = GREATEST(0, prihlasenych - 1) WHERE id = '$id'");
+					}
+				}
 				if ($result == FALSE)
 					die ("Nepodařilo se změnit přihlášku člena.");
-				if ($result !== false && mysqli_affected_rows($db_conn) > 0) {
-					query_db("UPDATE ".TBL_RACE." SET prihlasenych = GREATEST(0, prihlasenych - 1) WHERE id = '$id'");
-				}
 			}
 			else
 			{	// update
@@ -99,7 +107,12 @@ while ($zaznamZ=mysqli_fetch_array($vysledek))
 				$poz2=correct_sql_string($poz2);
 				$cterm=correct_sql_string($cterm);
 			
-				$result=query_db("UPDATE ".TBL_ZAVXUS." SET kat='$kat', pozn='$poz', pozn_in='$poz2', termin='$cterm', transport=$trans, sedadel=$sedl, ubytovani=$ubyt WHERE id_zavod = '$id' AND id_user = '$user'")
+				$sync_status_update = "";
+				if (!empty($zaznam_z['ext_id']) && $zaznamZ['sync_status'] !== 'PENDING_CREATE') {
+					$sync_status_update = ", sync_status='PENDING_UPDATE'";
+				}
+
+				$result=query_db("UPDATE ".TBL_ZAVXUS." SET kat='$kat', pozn='$poz', pozn_in='$poz2', termin='$cterm', transport=$trans, sedadel=$sedl, ubytovani=$ubyt".$sync_status_update." WHERE id_zavod = '$id' AND id_user = '$user'")
 					or die("Chyba při provádění dotazu do databáze.");
 				if ($result == FALSE)
 					die ("Nepodařilo se změnit přihlášku člena.");
@@ -115,7 +128,9 @@ while ($zaznamZ=mysqli_fetch_array($vysledek))
 				$poz2=correct_sql_string($poz2);
 				$cterm=correct_sql_string($cterm);
 			
-				$result=query_db("INSERT INTO ".TBL_ZAVXUS." (id_user, id_zavod, kat, pozn, pozn_in, termin, transport, sedadel,ubytovani) VALUES ('$user','$id','$kat','$poz','$poz2','$cterm',$trans,$sedl,$ubyt)")
+				$sync_status = !empty($zaznam_z['ext_id']) ? 'PENDING_CREATE' : 'LOCAL_ONLY';
+
+				$result=query_db("INSERT INTO ".TBL_ZAVXUS." (id_user, id_zavod, kat, pozn, pozn_in, termin, transport, sedadel,ubytovani, sync_status) VALUES ('$user','$id','$kat','$poz','$poz2','$cterm',$trans,$sedl,$ubyt,'$sync_status')")
 					or die("Chyba při provádění dotazu do databáze.");
 				if ($result == FALSE)
 					die ("Nepodařilo se změnit přihlášku člena.");
